stable json repository

+++ src/Repositories/BaseRepository.php
@@ -13,11 +13,41 @@ use FeatureKit\Support\DiscoverFeatures;
  */
 abstract class BaseRepository
 {
+    /**
+     * Final features output with FeatureKit\Factories\Feature instances.
+     *
+     * @var array
+     */
     protected array $features = [];
+
+    /**
+     * Features registered in a store
+     *
+     * @var array
+     */
     protected array $registered = [];
 
+    /**
+     * Load storage contents of features with array output.
+     *
+     * @return array
+     */
     abstract protected function loadRegisteredFeatures(): array;
+
+    /**
+     * Store an instance based on FeatureKit\Factories\Feature instance
+     *
+     * @param \FeatureKit\Factories\Feature $feature
+     * @return \FeatureKit\Factories\Feature
+     */
     abstract public function create(Feature $feature): Feature;
+
+    /**
+     * Delete an instance from storage based on FeatureKit\Factories\Feature instance
+     *
+     * @param \FeatureKit\Factories\Feature $feature
+     * @return \FeatureKit\Factories\Feature
+     */
     abstract public function delete(Feature $feature): Feature;
 
     public function __construct()
@@ -27,11 +57,9 @@ abstract class BaseRepository
                 if(!is_object($item)){
                     $item = (object) $item;
                 }
-
                 return [$item->key => $item];
             })->toArray();
         $this->analyzeFeatures();
-
     }
 
     private function analyzeFeatures(): void
@@ -47,16 +75,25 @@ abstract class BaseRepository
         }
 
         foreach($this->features as $key => $value){
-            if(!isset($instances[$key])){
+            if(empty($cache) && !isset($instances[$key])){
                 $this->delete($value);
+                unset($this->features[$key]);
             } else {
-                //foreach($instances[$key])
+                if(isset($this->registered[$key])){
+                    foreach($this->registered[$key] as $k => $v){
+                        $this->features[$key]->$k = $v;
+                    }
+                    $this->features[$key]->register();
+                }
             }
         }
-
-
     }
 
+    /**
+     * Get feature list from repository
+     *
+     * @return array
+     */
     public function getFeatures(): array
     {
         return $this->features;
